Fix moving some components to dedicated package
