+++ vector/src/main/java/im/vector/app/features/location/live/tracking/LocationSharingServiceConnection.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package im.vector.app.features.location
+package im.vector.app.features.location.live.tracking
 
 import android.content.ComponentName
 import android.content.Context
@@ -22,7 +22,6 @@ import android.content.Intent
 import android.content.ServiceConnection
 import android.os.IBinder
 import im.vector.app.core.di.ActiveSessionHolder
-import im.vector.app.features.location.live.tracking.LocationSharingAndroidService
 import im.vector.app.features.session.coroutineScope
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.launchIn
